fix: catgpt changes
