+++ package.json
@@ -24,7 +24,7 @@
   "dependencies": {
     "classnames": "^2.3.2",
     "daisyui": "^3.5.0",
-    "next": "^13.2.4",
+    "next": "^13.4.19",
     "next-seo": "^5.15.0",
     "next-sitemap": "^4.0.5",
     "react": "^18.2.0",
